refactor(BEB-33): move recursos de tipos de eventos para um diretorio proprio
