+++ src/modules/events/events-types/decorators/api-event-type-responses.decorator.ts
@@ -7,7 +7,7 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 
-import { EventTypeOutput } from '../../output';
+import { EventTypeOutput } from '../output';
 
 export function ApiCreateEventTypeResponses() {
   return applyDecorators(
