+++ src/modules/events/events-types/event-type.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { CreateEventTypeController } from '../controllers';
-import { EventTypeRepository } from '../repositories';
-import { CreateEventTypeService } from '../services';
+import { CreateEventTypeController } from './controllers';
 import { EventTypeEntity } from './entities';
+import { EventTypeRepository } from './repositories';
+import { CreateEventTypeService } from './services';
 
 @Module({
   imports: [TypeOrmModule.forFeature([EventTypeEntity])],
