+++ src/shared/auth/guard/index.ts
@@ -1,2 +1,3 @@
 export * from './admin.guard';
 export * from './auth.guard';
+export * from './user.guard';
